% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matching.R
\name{add_missing_salesforce}
\alias{add_missing_salesforce}
\title{Add missing companies}
\usage{
add_missing_salesforce(conn)
}
\arguments{
\item{conn}{connection to the anahita DB}
}
\description{
Add to the master table all companies that have
not been matched to CB.
}
\examples{
\dontrun{
conn <- connect_anahita()
add_missing_salesforce(conn)
}

}
