% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-domains.R
\name{assign_domain_type}
\alias{assign_domain_type}
\title{Stakeholder domain mapping}
\usage{
assign_domain_type(conn1, emails)
}
\arguments{
\item{conn1}{<- connect_anahita()}

\item{emails}{vector of emails}
}
\value{
a tibble
}
\description{
We're interested in mapping users into various
stakeholder groups. We can use various email domain
lists to do that.
}
\examples{
\dontrun{
conn1 <- connect_anahita()
emails <- c("ddd@edu.com", "ssss@gmail.com")
df <- assign_domain_type(conn1, dt)
}
}
