% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mixpanel.R
\name{build_advanced_analytics_table}
\alias{build_advanced_analytics_table}
\title{Advanced Analytics}
\usage{
build_advanced_analytics_table(events_si)
}
\arguments{
\item{events_si}{tibble of mixpanel events from serendipity (see \code{mxp_get_sry_table})}
}
\value{
a tibble
}
\description{
Event view for advanced analytics related events only
}
\examples{
\dontrun{
df <- build_advanced_analytics_table(events_si)
}

}
