% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mixpanel.R
\name{build_conversion_funnel}
\alias{build_conversion_funnel}
\title{Membership conversion tracking}
\usage{
build_conversion_funnel(stripe, events_si, events_tpk)
}
\arguments{
\item{stripe}{tibble of stripe customers (see \code{stripe_customers})}

\item{events_si}{tibble of mixpanel events from serendipity (see \code{mxp_get_sry_table})}

\item{events_tpk}{tibble of mixpanel events from toplink (see \code{mxp_get_sry_table})}
}
\description{
Membership conversion tracking
}
