% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mixpanel.R
\name{build_custom_map_table}
\alias{build_custom_map_table}
\title{Custom maps}
\usage{
build_custom_map_table(events_si)
}
\arguments{
\item{events_si}{tibble of mixpanel events from serendipity (see \code{mxp_get_sry_table})}
}
\value{
a tibble
}
\description{
Event view for custom map related events only
}
\examples{
\dontrun{
df <- build_custom_map_table(events_si)
}

}
