% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mixpanel.R
\name{build_insights_table}
\alias{build_insights_table}
\title{Insights}
\usage{
build_insights_table(events_si)
}
\arguments{
\item{events_si}{tibble of mixpanel events from serendipity (see \code{mxp_get_sry_table})}
}
\value{
a tibble
}
\description{
Event view for insights related events only
}
\examples{
\dontrun{
df <- build_insights_table(events_si)
}

}
