% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mixpanel.R
\name{build_issues_table}
\alias{build_issues_table}
\title{Issues}
\usage{
build_issues_table(events_si)
}
\arguments{
\item{events_si}{tibble of mixpanel events from serendipity (see \code{mxp_get_sry_table})}
}
\value{
a tibble
}
\description{
Event view for issues related events only
}
\examples{
\dontrun{
df <- build_issues_table(events_si)
}

}
