% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cb.R
\name{cb_download_zip}
\alias{cb_download_zip}
\title{Download CB data}
\usage{
cb_download_zip(token)
}
\arguments{
\item{token}{CB api token}
}
\description{
For our applications, it is more convenient to
apply a mass download of all CB files as CSV's on
a daily freqeuncy than built-in live API
calls. In the future, we would like to combine our
services with live API calls.
}
