% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cb.R
\name{cb_push_data}
\alias{cb_push_data}
\title{Push Crunchbase Data}
\usage{
cb_push_data(conn, dir)
}
\arguments{
\item{conn}{connection to the anahita DB}

\item{dir}{Temp directory as returned by \code{\link[=cb_download_zip]{cb_download_zip()}}}
}
\description{
Loop over dumped CB files and process data for writing in
remote DB.
}
