% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cb.R
\name{classify_industry_to_platforms}
\alias{classify_industry_to_platforms}
\title{Industry Classification}
\usage{
classify_industry_to_platforms(conn, data)
}
\arguments{
\item{conn}{connection to anahita}

\item{data}{The entire \code{cb_organizations} table}
}
\description{
For our membership tracking and development
strategies, it is extremely useful to map
companies to internal platforms. We martch platforms
to the industry categories provided by CB.
}
