% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_get}
\alias{config_get}
\title{Get Config Parameter}
\usage{
config_get(var)
}
\arguments{
\item{var}{config variable}
}
\description{
Special method to retrieve configuration parameter
from local hidden config file.
}
