% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-forum-agenda.R
\name{dump_all_agenda_blogs}
\alias{dump_all_agenda_blogs}
\title{Dump all Agenda Blogs}
\usage{
dump_all_agenda_blogs(conn, pages = NULL, n = 25, verbose = TRUE)
}
\arguments{
\item{conn}{connection to the data anahita DB}

\item{pages}{total pages to retrieve (default is NULL to retrieve all)}

\item{n}{how many articles per pages (default is 25)}

\item{verbose}{verbose message or not (default = TRUE)}
}
\description{
Run the agenda blog query function accross all pages and
overwrite the main table in anahita.
}
\examples{
\dontrun{
conn <- connect_anahita()
dump_all_agenda_blogs(conn)
}

}
