% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-en.R
\name{en_experts_by_insights}
\alias{en_experts_by_insights}
\title{Get Experts by Insights}
\usage{
en_experts_by_insights(conn, ins_id = NULL)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{ins_id}{vector saleforce insight area id}
}
\value{
tibble
}
\description{
Get all active expert network members with expertise
}
\examples{
\dontrun{
conn <- connect_dw()
en_experts_by_insights(conn)
}

}
