% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-en.R
\name{en_members}
\alias{en_members}
\title{Get Expert Network Members}
\usage{
en_members(conn)
}
\arguments{
\item{conn}{connection to the data warehouse}
}
\value{
tibble
}
\description{
Get all active expert network members
}
\examples{
\dontrun{
conn <- connect_dw()
en_members(conn)
}

}
