% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-en.R
\name{en_orgs}
\alias{en_orgs}
\title{Get Expert Network Organisations}
\usage{
en_orgs(conn, org_ids = NULL)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{org_ids}{vector of org ids (see \code{en_members})}
}
\value{
tibble
}
\description{
Expert Network members are in theory always affiliated with an
exisitng organisation in Salesforce. We use this primary organisation
to build the list of all organisation directly (sometimes it is their
only engagement point with the Forum) or indirectly involved with the
Expert Network.
}
\examples{
\dontrun{
conn <- connect_dw()
en_orgs(conn)
ens <- en_members(conn)
en_orgs(conn, ens$org_id)
}

}
