% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-events.R
\name{events_attendance}
\alias{events_attendance}
\title{Events attendance}
\usage{
events_attendance(conn1, sel_ids = NULL, max_date = NULL, sessions = NULL)
}
\arguments{
\item{conn1}{connection to the DW}

\item{sel_ids}{vector of ids}

\item{max_date}{when to start pulling}

\item{sessions}{session ids}
}
\value{
a tibble
}
\description{
Collecting event registration and attendance informations for
all digital members (paying and institutional)
}
\examples{
\dontrun{
con_dw <- connect_dw()
events_attendance(con_dw)
}

}
