% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-events.R
\name{events_webinar_attendance}
\alias{events_webinar_attendance}
\title{Webinars attendance}
\usage{
events_webinar_attendance(conn1, emails = NULL)
}
\arguments{
\item{conn1}{connection to the DW}

\item{emails}{vector of emails for filtering}
}
\value{
a tibble
}
\description{
Special webinars series were created as a a value proposition
to digital members and New Champions. Tracking engagement was
important but these webinar series did not last for long.
}
\examples{
\dontrun{
con_dw <- connect_dw()
events_webinar_attendance(con_dw)
}

}
