% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-forum-agenda.R
\name{format_agenda_blogs}
\alias{format_agenda_blogs}
\title{Forum Agenda articles list to df}
\usage{
format_agenda_blogs(list)
}
\arguments{
\item{list}{a list object to parse (see 'api_call_agenda_articles')}
}
\value{
a \code{tibble}
}
\description{
Forum Agenda articles list to df
}
