% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dms.R
\name{get_dms_companies}
\alias{get_dms_companies}
\title{Digital Members Companies}
\usage{
get_dms_companies(conn, tbl_dms)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{tbl_dms}{a dm tibble returned by \code{get_dms}}
}
\value{
a tibble
}
\description{
Many SI users are auto-upgraded to premium SI membership because
their email domain is whitelisted as belonging to a company with
an active Forum engagement. To keep the format of the table returned
by \code{get_dms} bindable with other people account methods, we provide
another utility function to retrieve the company.
}
\details{
Identify the DM company based on whitelsiting
}
\examples{
\dontrun{
conn <- connect_dw()
dms <- get_dms(conn)
get_dms_companies(conn, dms)
}

}
