% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dms.R
\name{get_dms_si_engagement}
\alias{get_dms_si_engagement}
\title{DMs through SI engagement}
\usage{
get_dms_si_engagement(conn)
}
\arguments{
\item{conn}{connection to the data warehouse}
}
\value{
a tibble
}
\description{
Some partners have officially chosen SI as
an engagement point. Their employees are auto-upgraded
auto-upgraded DMs in virtue of that.
}
\examples{
\dontrun{
conn <- connect_dw()
dms <- get_dms(conn)
get_dms_companies(conn, dms)
}

}
