% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-forum-agenda.R
\name{get_forum_agenda_blogs}
\alias{get_forum_agenda_blogs}
\title{Forum Agenda articles}
\usage{
get_forum_agenda_blogs(pages = 1, n = 25, verbose = TRUE)
}
\arguments{
\item{pages}{how many pages to paginate through (default is 1)}

\item{n}{how many articles per pages (default is 25)}

\item{verbose}{verbose message or not (default = TRUE)}
}
\value{
a tibble
}
\description{
It is of strategic interest for communities to
closely monitor their members contributions to various
editprial products of the Forum, such as Agenda blogs or
reports. We maintain a table with all those publications
with others with accounts in SF.
}
\examples{
df <- get_forum_agenda_blogs(5, 10)

}
