% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dms.R
\name{get_ptms}
\alias{get_ptms}
\title{Public Transformation Map Users}
\usage{
get_ptms(conn, org_ids = NULL, people_ids = NULL)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{org_ids}{list of active organisation ids (default = NULL)}

\item{people_ids}{vector of user SF ids you might want to exclude (default = NULL)}
}
\value{
a tibble
}
\description{
We define Public Transformation Map users as SI user with a Public subscription level. These
are mostly Lite account types but also a few consituents with no active positions.
}
\details{
Identify different groups of digital members.
}
\examples{
\dontrun{
conn <- connect_dw()
dt1 <- get_all_ncs(conn)
get_ptms(conn, dt1$org_id)
}

}
