% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sf.R
\name{get_sf_people}
\alias{get_sf_people}
\title{Retrieve SF people}
\usage{
get_sf_people(conn1, conn2)
}
\arguments{
\item{conn1}{connection to the data warehouse}

\item{conn2}{connection to anahita}
}
\value{
a tibble
}
\description{
Retrieve of salesforce people accounts and calculate
hash256 value based on SF id. This is used to analyse
encrypted SF ids sent by Serendipity.
}
\examples{
\dontrun{
conn1 <- connect_dw()
conn2 <- connect_anahita()
get_sf_people(conn1, conn2)
}

}
