% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dms.R
\name{get_si_signups}
\alias{get_si_signups}
\title{Signups through SI}
\usage{
get_si_signups(conn)
}
\arguments{
\item{conn}{connection to the data warehouse}
}
\value{
a tibble
}
\description{
We define the total number of Public SI users as the number of people
who signed-up to SI, irrespective of status and constituency type.
}
\details{
Build milestone metric aroudn signup
}
\examples{
\dontrun{
conn <- connect_dw()
dt1 <- get_si_signups(conn)
}

}
