% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-si-usage.R
\name{get_si_user_location}
\alias{get_si_user_location}
\title{Top SI user statistics - Location}
\usage{
get_si_user_location(
  conn,
  nmax = 5,
  id = NULL,
  time_since = NULL,
  time_to = NULL
)
}
\arguments{
\item{conn}{connection to the anahita DB}

\item{nmax}{max numbe of cities per user (default is 5)}

\item{id}{vector of SF id of individuals (default is NULL)}

\item{time_since}{starting time (default is NULL)}

\item{time_to}{ending time (default is NULL)}
}
\value{
a tibble
}
\description{
Calculating the table for all unique users is proven to be more efficient
than sending ids in the query. So we calculate these summary statistics on
a weekly basis.
}
\details{
Produce full table of key SI usage statistics for all si
unique users: top 5 briwasing cities with events.
}
\examples{
\dontrun{
conn <- connect_anahita()
get_si_user_location(conn, time_since = as.Date("2021-01-01"))
get_si_user_location(conn, nmax = 10, time_since = as.Date("2021-01-01"))
}

}
