% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-si-usage.R
\name{get_si_user_stats}
\alias{get_si_user_stats}
\title{SI user statistics - Total Activity}
\usage{
get_si_user_stats(conn, id = NULL, time_since = NULL, time_to = NULL)

get_si_user_stats(conn, id = NULL, time_since = NULL, time_to = NULL)
}
\arguments{
\item{conn}{connection to the anahita DB}

\item{id}{vector of SF id of individuals (default is NULL)}

\item{time_since}{starting time (default is NULL)}

\item{time_to}{ending time (default is NULL)}
}
\value{
a tibble

a tibble
}
\description{
Calculating the table for all unique users is proven to be more efficient
than sending ids in the query. So we calculate these summary statistics on
a weekly basis.

Calculating the table for all unique users is proven to be more efficient
than sending ids in the query. So we calculate these summary statistics on
a weekly basis.
}
\details{
Produce full table of key SI usage statistics for all si
unique users: total events, first_activity and last_activity

Produce full table of key SI usage statistics for all si
unique users: total events, first_activity and last_activity
}
\examples{
\dontrun{
conn <- connect_anahita()
get_si_user_stats(conn, time_since = as.Date("2021-01-01"))
}

\dontrun{
conn <- connect_anahita()
get_si_user_stats(conn, time_since = as.Date("2021-01-01"))
}

}
