% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-si-usage.R
\name{get_si_user_top_issues}
\alias{get_si_user_top_issues}
\title{Top Key Issues}
\usage{
get_si_user_top_issues(
  conn,
  nmax = 5,
  id = NULL,
  time_since = NULL,
  time_to = NULL
)
}
\arguments{
\item{conn}{connection to the anahita DB}

\item{nmax}{max numbe of insights per user (default is 5)}

\item{id}{vector of SF id of individuals (default is NULL)}

\item{time_since}{starting time (default is NULL)}

\item{time_to}{ending time (default is NULL)}
}
\value{
a tibble
}
\description{
Based on the results of \code{get_si_browsing} that aggregates all
browsing information into a summary table, we can quickly build
some key statistics around users. Here we calculate the top 5
key issues that users are interested in.
}
\details{
Produce top 5 key issues of interest for every
distinct user.
}
\examples{
\dontrun{
conn <- connect_anahita()
get_si_user_top_insights(conn, time_since = as.Date("2021-01-01"))
id <- "001b000003JXhbEAAT"
get_si_user_top_insights(conn, nmax = 10, id)
}

}
