% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tlo_communities}
\alias{get_tlo_communities}
\title{Get Organisations}
\usage{
get_tlo_communities(conn, community_ids)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{community_ids}{vector of community ids}
}
\value{
tibble
}
\description{
Contract based communities are automatically synced with contract information.
However, some communities are not contract based and should be accessed by querying
directly the community.
}
\details{
Get list of organisations based on membership affiliation
}
\examples{
\dontrun{
conn <- connect_dw()
get_tlo_communities(conn, "New Champion")
}

}
