% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tlo_contacts}
\alias{get_tlo_contacts}
\title{Get TFE and OPS Contacts}
\usage{
get_tlo_contacts(conn, org_ids)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{org_ids}{list of organisation ids}
}
\value{
tibble
}
\description{
One needs to provide both organisation ids and names
because the main accounts table contains organisation ids but only
tlo names.
}
\details{
Get list of of top level and organisational contacts for a list
of organisations.
}
\examples{
\dontrun{
conn <- connect_dw()
dt <- get_all_ncs(conn)
get_tlo_contacts(conn, dt$org_id)
}

}
