% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tlo_contracts}
\alias{get_tlo_contracts}
\title{Get Organisations}
\usage{
get_tlo_contracts(conn, engagement_levels, membership_types)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{engagement_levels}{vector that defines the contract engagement type}

\item{membership_types}{vector that defines the contract engagement type}
}
\value{
tibble
}
\description{
For safety, one is asked to provide both the engagement level and the the
contract membership type. These can be reviewed with the method \code{get_contract_membership_types}
}
\details{
Get list of organisations based on contract characteristics.
}
\examples{
\dontrun{
conn <- connect_dw()
get_tlo_contracts(conn, "New Champion")
}

}
