% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tlo_invoices}
\alias{get_tlo_invoices}
\title{Get invoices}
\usage{
get_tlo_invoices(conn, org_ids)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{org_ids}{list of organisation ids}
}
\value{
a tibble
}
\description{
Get list of invoices for organisations
}
\examples{
\dontrun{
conn <- connect_dw()
dt <- get_all_ncs4(conn)
get_tlo_invoices(conn, dt$org_id)
}

}
