% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tlo_opportunities}
\alias{get_tlo_opportunities}
\title{Get Opportunities}
\usage{
get_tlo_opportunities(conn, org_ids)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{org_ids}{list of organisation ids}
}
\value{
tibble
}
\description{
Before engaging with a company or communication about
engagement opportunities, it is useful to check existing
opportunities. We select here the following opportunity
types: "Approved", "Closing", "Finalist", "Idea", "Proposal",
and "Qualification"
}
\details{
Get the full list of running TLO opportunities
}
\examples{
\dontrun{
conn <- connect_dw()
dt <- get_all_ncs(conn)
get_tlo_opportunities(conn, dt$org_id)
}

}
