% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-whitelisting.R
\name{get_whitelisted_domains}
\alias{get_whitelisted_domains}
\title{Get whitelisted domains}
\usage{
get_whitelisted_domains(conn, org_ids = NULL)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{org_ids}{list of organisation ids}
}
\value{
a tibble
}
\description{
Get list all whitelisted domains in Salesforce
or domains whitelisted for a selected organisation.
}
\examples{
\dontrun{
conn <- connect_dw()
get_whitelisted_domains(conn)
get_whitelisted_domains(conn, "001b0000005gxmlAAA")
}

}
