% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gfcs.R
\name{gfc_members}
\alias{gfc_members}
\title{Get Global Future Councils}
\usage{
gfc_members(conn)
}
\arguments{
\item{conn}{connection to the data warehouse}
}
\value{
tibble
}
\description{
Get all active GFC members
}
\examples{
\dontrun{
conn <- connect_dw()
gfc_members(conn)
}

}
