% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gis.R
\name{gis_all_orgs}
\alias{gis_all_orgs}
\title{Retrieve all GIs}
\usage{
gis_all_orgs(conn, only_active = TRUE)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{only_active}{select only contracts that are "Activated/Running" (default is TRUE)}
}
\value{
a tibble
}
\description{
Retrieve all GI companies based on contract
data.
}
\examples{
\dontrun{
conn <- connect_dw()
gis_all_orgs(conn)
}

}
