% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gis.R
\name{gis_all_orgs_by_membership}
\alias{gis_all_orgs_by_membership}
\title{Retrieve all GI companies}
\usage{
gis_all_orgs_by_membership(conn)
}
\arguments{
\item{conn}{connection to the data warehouse}
}
\value{
a tibble
}
\description{
Retrieving GI companies is done pulling from three different
communities in SF. GI here includes Technolohy Pioneers.
}
\details{
Retrieve all companies within the Global Innovators
SF community.
}
\examples{
\dontrun{
conn <- connect_dw()
gis_all_orgs_by_membership(conn)
}

}
