% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gis.R
\name{gis_all_orgs_gis}
\alias{gis_all_orgs_gis}
\title{Retrieve all GI companies}
\usage{
gis_all_orgs_gis(conn, only_active = TRUE)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{only_active}{select only contracts that are "Activated/Running" (default is TRUE)}
}
\value{
a tibble
}
\description{
Retrieve all GI C4IR companies based on
contract information. At some point all C4IR
contracts will move to the GI engagement level.
}
\details{
Retrieve all GI companies based on
contract information.
}
\examples{
\dontrun{
conn <- connect_dw()
gis_all_orgs_gis(conn)
}

}
