% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gis.R
\name{gis_contacts}
\alias{gis_contacts}
\title{Retrieve all GI contacts}
\usage{
gis_contacts(conn, org_ids = NULL)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{org_ids}{list of organisation ids (calculated if not provided)}
}
\value{
a tibble
}
\description{
Retrieve all contacts from the GI
companies.
}
\examples{
\dontrun{
conn <- connect_dw()
dt <- gis_all_orgs(conn)
gis_contacts(conn, dt$membername__c)
}

}
