% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gis.R
\name{gis_invoices}
\alias{gis_invoices}
\title{Get Org Invoices}
\usage{
gis_invoices(conn, org_ids = NULL)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{org_ids}{list of organisation ids (default NULL)}
}
\value{
a tibble
}
\description{
Retrieve list of invoices from organisations.
}
\examples{
\dontrun{
conn <- connect_dw()
dt <- gis_all_orgs(conn)
gis_invoices(conn)
gis_invoices(conn, dt$org_id)
}

}
