% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gis.R
\name{gis_members}
\alias{gis_members}
\title{Retrieve all GI consituents}
\usage{
gis_members(conn, org_ids = NULL)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{org_ids}{list of organisation ids (if not provided will run \code{gis_all_orgs})}
}
\value{
a tibble
}
\description{
We simply retrieve all active accounts for which the
primary organisation is a GI.
}
\details{
Retrieve all constituents from the GI
companies.
}
\examples{
\dontrun{
conn <- connect_dw()
dt <- gis_all_orgs(conn)
gis_members(conn, dt$membername__c)
}

}
