% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gis.R
\name{gis_opportunities}
\alias{gis_opportunities}
\title{Get Org Opportunities}
\usage{
gis_opportunities(conn, org_ids = NULL)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{org_ids}{list of organisation ids (default NULL)}
}
\value{
a tibble
}
\description{
Before engaging with a potential company for a contract renewal or
new engagement opportunity, it is useful to check whether that
company has any formal conversations with other teams already. We select
opportunities with status "Approved", "Closing", "Finalist", "Idea", "Proposal",
and "Qualification"
}
\details{
Retrieve list of "open" opportunities.
}
\examples{
\dontrun{
conn <- connect_dw()
dt <- gis_all_orgs(conn)
gis_opportunities(conn)
gis_opportunities(conn, dt$org_id)
}

}
