% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gis.R
\name{gis_org_pems}
\alias{gis_org_pems}
\title{Get PEMs}
\usage{
gis_org_pems(conn, dt = NULL)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{dt}{list of companies to speed up calculations (see \code{gis_all_orgs})}
}
\value{
a tibble
}
\description{
Getting a meaningful PEM list for a list of organisations is slightly
convoluted for NCs as you need to look both at the subsidary and tlo
level.
}
\details{
Build the full list of primary and secondary PEMS.
}
\examples{
\dontrun{
conn <- connect_dw()
dt <- gis_all_orgs(conn)
gis_org_pems(conn)
gis_org_pems(conn, dt)
}

}
