% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gis.R
\name{gis_ptms}
\alias{gis_ptms}
\title{Retrieve all GI Digital Members}
\usage{
gis_ptms(conn, org_ids, people_ids = NULL)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{org_ids}{list of organisation ids (if not provided will run \code{gis_all_orgs})}

\item{people_ids}{list of people ids see \code{nc_members} and \code{nc_constituents} (default is NULL)}
}
\value{
a tibble
}
\description{
SI users with a domain found to match the domain of a
GI company are automatically upgraded to a premium
membership which grants them access to open events and other
things. Because coding is not 100\% clean in SF, we provide an
exclusion list of ids for accounts already identified to be
members or consituents for NC companies. This should be
simplified in the future.
}
\details{
Retrieve all PTMs from the GI companies.
}
\examples{
\dontrun{
conn <- connect_dw()
dt <- gis_all_orgs(conn)
gis_members(conn, dt$membername__c)
}

}
