% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-mixpanel.R
\name{initialize_all_monthly_views}
\alias{initialize_all_monthly_views}
\title{Initialize monthly views}
\usage{
initialize_all_monthly_views(dev = config_get(deploy_status))
}
\arguments{
\item{dev}{deployment status (dev does not push tables)}
}
\value{
a tibble
}
\description{
Start monthly views from scratch using a time
query to mixpanel events table.
}
\examples{
\dontrun{
initialize_all_monthly_views("dev")
}

}
