% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-events.R
\name{initialize_events_tables}
\alias{initialize_events_tables}
\title{Initialize events table for DMS}
\usage{
initialize_events_tables(dev = config_get(deploy_status))
}
\arguments{
\item{dev}{deployment status read from config (dev will not right into DBs)}
}
\value{
list of tibbles
}
\description{
Certain events being value proposition for DMS, we track
actively their registration and participation.
}
\examples{
\dontrun{
initialize_events_tables()
}

}
