% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-knowledge.R
\name{knwl_outer_ring_stats}
\alias{knwl_outer_ring_stats}
\title{Build outer ring article volume co-occurence}
\usage{
knwl_outer_ring_stats(conn1, time_since = NULL, time_to = NULL)
}
\arguments{
\item{conn1}{connection to anahita}

\item{time_since}{start time (default NULL)}

\item{time_to}{end time (default NULL)}
}
\description{
Build outer ring article volume co-occurence
}
