% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-knowledge.R
\name{knwl_toplink_network}
\alias{knwl_toplink_network}
\title{Save TM network}
\usage{
knwl_toplink_network()
}
\value{
a tibble
}
\description{
Salesforce not being a graph DB, it is complicated to
rebuild the full TM network in a friendly format. The
Toplink endpoint allows to do just that.
}
