% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-knowledge.R
\name{knwl_update_articles}
\alias{knwl_update_articles}
\title{Update knowledge articles}
\usage{
knwl_update_articles(
  dev = config_get(deploy_status),
  envr = "QA",
  time_since = NULL,
  api_token = config_get(api_key_neo4j_qa)
)
}
\arguments{
\item{dev}{deployment status read from config (dev will not right into DBs)}

\item{envr}{environment (PROD, QA, STAGING)}

\item{time_since}{starting time for last modification article time}

\item{api_token}{knowledge endpoint api token}
}
\value{
a list of tibbles
}
\description{
We populated a table of articles processed by Serendipity in their
backennd. It served to us by a endpoint maintained by Serendipity.
Provides many information related to moderations and classification.
}
\examples{
\dontrun{
knwl_update_articles("dev", "QA")
}

}
