% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-knowledge.R
\name{knwl_update_outer_ring_stats}
\alias{knwl_update_outer_ring_stats}
\title{Outer ring stats}
\usage{
knwl_update_outer_ring_stats(
  dev = config_get(deploy_status),
  time_since = Sys.time() - 3600 * 24 * 365,
  time_to = NULL
)
}
\arguments{
\item{dev}{deployment status read from config (dev will not right into DBs)}

\item{time_since}{start time (default NULL)}

\item{time_to}{end time (default is one year back)}
}
\value{
a tibble
}
\description{
We calculate the volume of articles co-occuring accross pairs
of topics.
}
\examples{
\dontrun{
con_an <- connect_anahita()
dt <- knwl_update_weekly_moderation_targets(con_an)
}

}
