% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-knowledge.R
\name{knwl_update_sources}
\alias{knwl_update_sources}
\title{Update knowledge articles}
\usage{
knwl_update_sources(
  dev = config_get(deploy_status),
  api_token = config_get("api_key_neo4j")
)
}
\arguments{
\item{dev}{deployment status read from config (dev will not right into DBs)}

\item{api_token}{knowledge endpoint api token}
}
\value{
a list of tibbles
}
\description{
We populated a table of articles processed by Serendipity in their
backennd. It served to us by a endpoint maintained by Serendipity.
Provides many information related to moderations and classification.
}
\examples{
\dontrun{
knwl_update_sources("dev")
}

}
