% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-knowledge.R
\name{knwl_update_tpk_network}
\alias{knwl_update_tpk_network}
\title{Weekly moderation stats}
\usage{
knwl_update_tpk_network(dev = config_get(deploy_status), dirpath = NULL)
}
\arguments{
\item{dev}{deployment status read from config (dev will not right into DBs)}

\item{dirpath}{path to save local copy of file}
}
\value{
a tibble
}
\description{
Weekly moderation targets are based on achieving a number of
targets based on total volume of articles but also number of
approved articles per topic and diversity of sources per topic
among the approved articles.
}
\examples{
\dontrun{
con_an <- connect_anahita()
dt <- knwl_update_tpk_network("./")
}

}
