% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matching.R
\name{match_to_domain}
\alias{match_to_domain}
\title{Matching Strengt 2}
\usage{
match_to_domain(conn)
}
\arguments{
\item{conn}{connection to the anahita DB}
}
\description{
Domain matching
}
\examples{
\dontrun{
conn <- connect_anahita()
match_to_domain(conn)
}

}
