% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matching.R
\name{match_to_legal_name}
\alias{match_to_legal_name}
\title{Matching Strengt 3}
\usage{
match_to_legal_name(conn)
}
\arguments{
\item{conn}{connection to the anahita DB}
}
\description{
Legal name matching
}
\examples{
\dontrun{
conn <- connect_anahita()
match_to_legal_name(conn)
}

}
