% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mixpanel.R
\name{mxp_clean_events}
\alias{mxp_clean_events}
\title{Clean mixpanel events}
\usage{
mxp_clean_events(conn1, dt, email_to_ids, source = "sry")
}
\arguments{
\item{conn1}{a connection to anahita DB}

\item{dt}{mixpanel events tibble (see \code{mxp_get_sry_table} or \code{mxp_get_tpk_table})}

\item{email_to_ids}{tibble of email to id mappings (see \code{sf_build_user_emails_to_id})}

\item{source}{mixpanel source, Serendipity (sry) or other (tpk e.g.)}
}
\value{
a tibble
}
\description{
Raw mixpanel event suffer issues in tracking
implemention that need to be fixed for cleaner data.
}
\examples{
\dontrun{
conn1 <- connect_anahita()
df <- mxp_clean_events(conn1)
}

}
