% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mixpanel.R
\name{mxp_fetch_data}
\alias{mxp_fetch_data}
\title{Fetch mixpanel data}
\usage{
mxp_fetch_data(from_date, to_date, api_secret, col_sels, dir)
}
\arguments{
\item{from_date}{start date}

\item{to_date}{end date}

\item{api_secret}{mixpanel API}

\item{col_sels}{vector of columns to select}

\item{dir}{temp dir location}
}
\value{
a tibble
}
\description{
Will use the mixpanel export API to dump files in local temp dir
and then read back in the JSON.
}
\examples{
\dontrun{
t1 = Sys.Date() - 1
t2 = Sys.Date() - 1
api <- config_get(mp_si)
cols <- c("time", "distinct_id")
dir <- "./mpdata"
mxp_set_up_dir(dir)
df <- mxp_fetch_data(t1, t2, api, cols, dir)
}

}
