% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mixpanel.R
\name{mxp_fix_earthtime_story_names}
\alias{mxp_fix_earthtime_story_names}
\title{Fix earthtime story names}
\usage{
mxp_fix_earthtime_story_names(conn1, dt)
}
\arguments{
\item{conn1}{connection to anahita DB}

\item{dt}{a mixpanel events table}
}
\value{
a tibble
}
\description{
Earthtime events are not properly recognized. We need to
use story names to tag these events as earthtime events
for easier manipulation.
}
\examples{
\dontrun{
conn1 <- connect_anahita()
dt <- fetch_mixpanel_data()
df <- mxp_fix_earthtime_story_names(conn1, dt)
}

}
