% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mixpanel.R
\name{mxp_fix_ids}
\alias{mxp_fix_ids}
\title{Replace emails by ids}
\usage{
mxp_fix_ids(dt, tbl_users)
}
\arguments{
\item{dt}{a mixpanel events table}

\item{tbl_users}{table of emails to id mapping (see \code{mxp_build_user_ids_emails})}
}
\value{
a tibble
}
\description{
Utility function that replaces email identifiers by
their associated SF id.
}
\examples{
\dontrun{
conn1 <- connect_anahita()
dt1 <- fetch_mixpanel_data()
dt2 <- mxp_build_user_ids_emails(conn1)
df <- mxp_fix_ids(dt1, dt2)
}

}
