% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mixpanel.R
\name{mxp_fix_topic_ids}
\alias{mxp_fix_topic_ids}
\title{Enrich topic ids}
\usage{
mxp_fix_topic_ids(dt)
}
\arguments{
\item{dt}{a mixpanel events table}
}
\value{
a tibble
}
\description{
Certain events can be associated to browsing a given topic
but mixpanel will register the id related to the specific event.
We create a new topic id to enrich with the topic id found in the
URL.
}
\examples{
\dontrun{
dt <- fetch_mixpanel_data()
df <- mxp_fix_topic_ids(dt)
}

}
