% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mixpanel.R
\name{mxp_get_tpk_table}
\alias{mxp_get_tpk_table}
\title{Toplink Mixpanel Events}
\usage{
mxp_get_tpk_table(conn1, fpath)
}
\arguments{
\item{conn1}{a connection to anahita DB}

\item{fpath}{path to dumped file if exsists}
}
\value{
a tibble
}
\description{
Load or dump events data. The dump could last several
hours.
}
\examples{
\dontrun{
conn1 <- connect_anahita()
df <- mxp_get_tpk_table(conn1)
}

}
