% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mixpanel.R
\name{mxp_track_weforum_logins}
\alias{mxp_track_weforum_logins}
\title{SI logins from Forum domain}
\usage{
mxp_track_weforum_logins(events_si)
}
\arguments{
\item{events_si}{tibble of mixpanel events from serendipity (see \code{mxp_get_sry_table})}
}
\description{
We track all signups that orginate from a weforum domain referral.
We also track all signups from weforum domains as one measure of
retention.
}
