% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ncs.R
\name{ncs_all_fme_contracts}
\alias{ncs_all_fme_contracts}
\title{Retrieve all FME contracts}
\usage{
ncs_all_fme_contracts(conn)
}
\arguments{
\item{conn}{connection to the data warehouse}
}
\value{
a tibble
}
\description{
Retrieve all companies with a Forum Memberhip contracts. We
keep all contract status for historical analysis.
}
\examples{
\dontrun{
conn <- connect_dw()
ncs_all_fme_contracts(conn)
}

}
