% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ncs.R
\name{ncs_all_nc_contracts}
\alias{ncs_all_nc_contracts}
\title{Retrieve all NC contracts}
\usage{
ncs_all_nc_contracts(conn)
}
\arguments{
\item{conn}{connection to the data warehouse}
}
\value{
a tibble
}
\description{
Retrieve all companies with a New Champions contract. We
keep all contract status for historical analysis.
}
\examples{
\dontrun{
conn <- connect_dw()
var_sel <- c("accountid","top_level_organization__c")
ncs_all_nc_contracts(conn)
}

}
