% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ncs.R
\name{ncs_all_orgs}
\alias{ncs_all_orgs}
\title{Retrieve all NC contracts (NCs + FMEs)}
\usage{
ncs_all_orgs(conn, only_active = TRUE)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{only_active}{select only contracts that are "Activated/Running" (default is TRUE)}
}
\value{
a tibble
}
\description{
Retrieve all companies with a New Champions or FME contract. We
keep all contract status for historical analysis.
}
\examples{
\dontrun{
conn <- connect_dw()
ncs_all_orgs(conn)
}

}
