% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ncs.R
\name{ncs_all_orgs_by_membership}
\alias{ncs_all_orgs_by_membership}
\title{Retrieve all NC companies}
\usage{
ncs_all_orgs_by_membership(conn)
}
\arguments{
\item{conn}{connection to the data warehouse}
}
\value{
a tibble
}
\description{
Retrieve all NC companies pulling from the contract
based communities.
}
\examples{
\dontrun{
conn <- connect_dw()
ncs_all_orgs_by_membership(conn)
}

}
