% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ncs.R
\name{ncs_check_missing_companies}
\alias{ncs_check_missing_companies}
\title{Missing companies}
\usage{
ncs_check_missing_companies(conn)
}
\arguments{
\item{conn}{connection to the data warehouse}
}
\value{
a named list
}
\description{
Check if companies are missing or need to be
added in the NC contract base communities.
}
\examples{
\dontrun{
conn <- connect_dw()
dt <- ncs_all_orgs(conn)
to_add <- dt$to_add
to_remove <- dt$to_remove
}

}
