% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ncs.R
\name{ncs_contacts}
\alias{ncs_contacts}
\title{Get Contacts}
\usage{
ncs_contacts(conn, org_ids = NULL)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{org_ids}{list of organisation ids (default NULL)}
}
\value{
a tibble
}
\description{
Retrieve TFE and Operational contacts for NC companies.
}
\examples{
\dontrun{
conn <- connect_dw()
dt <- ncs_all_orgs(conn)
ncs_contacts(conn)
ncs_contacts(conn, dt$org_id)
}

}
