% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ncs.R
\name{ncs_dms}
\alias{ncs_dms}
\title{NC DM accounts}
\usage{
ncs_dms(conn, org_ids, people_ids = NULL)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{org_ids}{list of active organisation ids}

\item{people_ids}{list of people ids see \code{ncs_members} and \code{ncs_constituents} (default is NULL)}
}
\value{
a tibble
}
\description{
SI users with a domain found to match the domain of a
New Champion company are automatically upgraded to a premium
membership which grants them access to open events and other
things. Because coding is not 100\% clean in SF, we provide an
exclusion list of ids for accounts already identified to be
members or consituents for NC companies. This should be
simplified in the future. The function uses the whitelisting to
overwrite missing primary organisation ids with the organisation
associated to the domain in the whitelisting table.
}
\details{
Identify digital members with domains associated to
New Champion Companies.
}
\examples{
\dontrun{
conn <- connect_dw()
dt1 <- ncs_all_orgs(conn)
list_ncs_members <- ncs_members(conn, "a0e0X00000hzdENQAY")
list_ncs_constituents <- ncs_constituents(conn, dt1$org_id)
people_ids <- unique(c(list_ncs_members$id, list_ncs_constituents$id))
ncs_dms(conn, dt1$org_id, people_ids)
}

}
