% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ncs.R
\name{ncs_initialize_tlo_quarterly_anniversaries}
\alias{ncs_initialize_tlo_quarterly_anniversaries}
\title{NC anniversaries}
\usage{
ncs_initialize_tlo_quarterly_anniversaries(conn, dt_ncs = NULL)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{dt_ncs}{tibble of all active nc companies (see \code{ncs_all_orgs})}
}
\value{
a tibble
}
\description{
Utility function to determine the quarterly anniveraires for
active NC companies. This table is at the basis of the quarterly
communication app build for the NC team.
}
\details{
Calculate NC quarterly anniversiaries table.
}
\examples{
\dontrun{
conn <- connect_dw()
dt <- ncs_all_orgs(conn)
ncs_initialize_tlo_quarterly_anniversaries(conn, dt)
}

}
