% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ncs.R
\name{ncs_members}
\alias{ncs_members}
\title{Get people with NC membership}
\usage{
ncs_members(conn, membership_ids, org_ids = NULL)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{membership_ids}{list of membership ids}

\item{org_ids}{list of organisation ids (default NULL)}
}
\value{
a tibble
}
\description{
Because there is no automation to remove people from
the community when a contract is terminated, you can
specify the list of active organisation ids to make sure
you reflect number for active companies.
}
\details{
Retrieve all active memberships in the New
Champions constituent community.
}
\examples{
\dontrun{
conn <- connect_dw()
dt <- ncs_all_orgs(conn)
ncs_members(conn, "a0e0X00000hzdENQAY")
ncs_members(conn, "a0e0X00000hzdENQAY", dt$org_id)
}

}
