% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ncs.R
\name{ncs_ptms}
\alias{ncs_ptms}
\title{NC PTM accounts}
\usage{
ncs_ptms(conn, org_ids, people_ids = NULL)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{org_ids}{list of active organisation ids}

\item{people_ids}{list of people ids see \code{ncs_members} and \code{ncs_constituents} (default is NULL)}
}
\value{
a tibble
}
\description{
By defintion, if the whitelisted domains for our NC
companies are complete, there should be no accounts with
a public subscription.
}
\details{
Identify public transformation map users.
}
\examples{
\dontrun{
conn <- connect_dw()
dt <- ncs_all_orgs(conn)
ncs_ptms(conn, dt$org_id)
}

}
