% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{org_members}
\alias{org_members}
\title{Get org constituents}
\usage{
org_members(conn, org_id, tlo_only = FALSE)
}
\arguments{
\item{conn}{connection to the data warehouse}

\item{org_id}{organisation id}

\item{tlo_only}{only tlo members rather than all subsidaries (default is FALSE)}
}
\value{
a tibble
}
\description{
Due to the structure of organisations in SF, one can
retrieve the accounts explicitly coded under a given
organisation or all those that fall under each subsidary.
}
\details{
Retrieve all constituents coded under a given
organisation.
}
\examples{
\dontrun{
conn <- connect_dw()
org_members(conn, "a0e0X00000hzdENQAY")
}

}
