% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sf.R
\name{sf_build_user_emails_to_id}
\alias{sf_build_user_emails_to_id}
\title{User ids and email list}
\usage{
sf_build_user_emails_to_id(conn1)
}
\arguments{
\item{conn1}{connection to the anahita DB}
}
\value{
a tibble
}
\description{
Sometimes, mixpanel stores emails instead of SF ids. We need
to replace emails by ids using a reference list of email to id
mappings. We need to note more than 1 email can be associated to
an single id since we have the person emnail and the toplink email.
In principle, the toplink email is the right one but we never now.
}
\examples{
\dontrun{
conn1 <- connect_anahita()
df <- mxp_build_user_ids_emails(conn1)
}

}
