% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-si-usage.R
\name{si_usage_assign_dm_member_type}
\alias{si_usage_assign_dm_member_type}
\title{Assign category}
\usage{
si_usage_assign_dm_member_type(
  record_type_id,
  si_sub_type,
  si_membership,
  stripe_status
)
}
\arguments{
\item{record_type_id}{account recort type id}

\item{si_sub_type}{digital membership account categorization}

\item{si_membership}{membership level (Public, Premium, Pro, Platinum)}

\item{stripe_status}{stripe status plan}
}
\value{
a character string
}
\description{
For business reporting purposes, we want to assign every individual
SI user a category reflecting its digital membership status such as
being a paying membership or a whitelisted digital member.
}
\details{
Assign category to SI users.
}
\examples{
\dontrun{
record_type_id <- "012b00000000ARJAA2"
si_sub_type <- "TopLink"
si_membership <- "Pro"
stripe_status <- "active"
si_usage_assign_dm_member_type(record_type, si_sub_type, si_membership, stripe_status)
}

}
