% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-si-usage.R
\name{si_usage_assign_dm_record_type}
\alias{si_usage_assign_dm_record_type}
\title{Assign category}
\usage{
si_usage_assign_dm_record_type(record_type_id)
}
\arguments{
\item{record_type_id}{vector of account recort type id}
}
\value{
a character string
}
\description{
For business reporting purposes, we want to assign every individual
SI user to high level category based on the SF account type.
}
\details{
Assign category to SI users.
}
\examples{
\dontrun{
si_usage_assign_dm_record_type(c("012b00000000ARJAA2", "012b00000000ARJAA2"))
}

}
