% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-si-usage.R
\name{si_usage_assign_email_type}
\alias{si_usage_assign_email_type}
\title{Email type}
\usage{
si_usage_assign_email_type(conn, emails)
}
\arguments{
\item{conn}{connection to anahita DB}

\item{emails}{vector of emails or domains}
}
\value{
a vector
}
\description{
We try to infer the user's stakeholder group by looking at the
email domain and or the extension and see if it matches any of the
reference lists we have.
}
\details{
User email type
}
\examples{
\dontrun{
conn <- connect_anahita()
emails <- c("ss@dd.com", "ddd.fr", "dsdsd@hrd.de")
si_usage_assign_email_type(conn, emails)
}

}
