% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-si-usage.R
\name{si_usage_assign_grouped_domain}
\alias{si_usage_assign_grouped_domain}
\title{Grouped domain}
\usage{
si_usage_assign_grouped_domain(conn, emails)
}
\arguments{
\item{conn}{connection to anahita DB}

\item{emails}{vector of emails or domains}
}
\value{
a vector
}
\description{
We assign to each emnail an domain group coinciding with the way
we build grouped domains based on SF TLO versus subsidary
organisations.
}
\details{
User grouped domain affiliation
}
\examples{
\dontrun{
conn <- connect_anahita()
emails <- c("ss@dd.com", "ddd.fr", "dsdsd@hrd.de")
si_usage_assign_grouped_domain(conn, emails)
}

}
