% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-si-usage.R
\name{si_usage_assign_org}
\alias{si_usage_assign_org}
\title{Assign Organisation}
\usage{
si_usage_assign_org(conn, dt)
}
\arguments{
\item{conn}{connection to anahita}

\item{dt}{accounts data frame}
}
\value{
a tibble
}
\description{
For business reporting purposes, we want to assign every individual
SI user an guess on their affiliated Top Level Organisation in SF, if
possible based on the email domain or SF account.
}
\details{
Assign organisation to SI users.
}
\examples{
\dontrun{
conn1 <- connect_dw()
conn2 <- connect_anahita()
accounts <- tbl(conn1, dbplyr::in_schema("ods","sfdc_account_v")) \%>\% 
 filter(si_sub_final_text__c == "Pro") \%>\%
 collect()
assigned_tlos <- si_usage_assign_tlo(conn2, accounts)
}

}
