% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-si-usage.R
\name{si_usage_extract_domain}
\alias{si_usage_extract_domain}
\title{Extract domain}
\usage{
si_usage_extract_domain(emails)
}
\arguments{
\item{emails}{vector of emails}
}
\value{
a vector
}
\description{
The email domain is used to derive information on the
user's stakeholder group (professional versus academic e.g.)
and affiliation to a know company (based on SF or CB domain
mapping)
}
\details{
Extract email domain
}
\examples{
\dontrun{
emails <- c("ss@dd.com", "test@ddd.fr", "dsdsd@hrd.de")
si_usage_extract_domain(emails)
}

}
