% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-si-usage.R
\name{si_usage_extract_extension}
\alias{si_usage_extract_extension}
\title{Extract extension}
\usage{
si_usage_extract_extension(emails)
}
\arguments{
\item{emails}{vector of emails or domains}
}
\value{
a vector
}
\description{
A unique company domain can be declined with several
extension identifying for example the country.
}
\details{
Extract domnain extension
}
\examples{
\dontrun{
emails <- c("ss@dd.com", "ddd.fr", "dsdsd@hrd.de")
si_usage_extract_domain(emails)
}

}
