% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stripe.R
\name{stripe_abandonned_payments}
\alias{stripe_abandonned_payments}
\title{Retrieve abandonned payments}
\usage{
stripe_abandonned_payments(conn)
}
\arguments{
\item{conn}{connection to the anahita Data Base}
}
\value{
tibble
}
\description{
Retrieve abandonned payments
}
\examples{
\dontrun{
conn <- connect_anahita()
stripe_abandonned_payments(conn)
}

}
