% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stripe.R
\name{stripe_customers}
\alias{stripe_customers}
\title{Retrieve Stripe Customers}
\usage{
stripe_customers(conn)
}
\arguments{
\item{conn}{connection to the anahita Data Base}
}
\value{
tibble
}
\description{
Retrieve Stripe Customers
}
\examples{
\dontrun{
conn <- connect_anahita()
stripe_customers(conn)
}

}
