% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stripe.R
\name{stripe_fetch_data}
\alias{stripe_fetch_data}
\title{Fetch Stripe Customers}
\usage{
stripe_fetch_data(stripe_key)
}
\arguments{
\item{stripe_key}{stripe API key}
}
\value{
tibble
}
\description{
Fetch Stripe Customers
}
\examples{
\dontrun{
stripe_key = config_get(stripe)
stripe_fetch_data(stripe_key)
}

}
