% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stripe.R
\name{stripe_stats_for_elvin}
\alias{stripe_stats_for_elvin}
\title{Business performance reporting}
\usage{
stripe_stats_for_elvin(conn1, conn2)
}
\arguments{
\item{conn1}{connection anahita}

\item{conn2}{connection dw}
}
\value{
tibble
}
\description{
Business performance reporting
}
\examples{
\dontrun{
conn1 <- connect_anahita()
conn2 <- connect_dw()
stripe_stats_for_elvin(stripe_key)
}

}
