% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-forum-agenda.R
\name{update_agenda_blogs}
\alias{update_agenda_blogs}
\title{Update Agenda Blogs}
\usage{
update_agenda_blogs(conn, pages = 5, n = 25, verbose = TRUE)
}
\arguments{
\item{conn}{connection to the data anahita DB}

\item{pages}{how far back you want to search for new articles (default is 5)}

\item{n}{how many articles per pages (default is 25)}

\item{verbose}{verbose message or not (default = TRUE)}
}
\description{
Since the endpoints provide a pagination approach but
results can't ne queries by time, the updates pulls a
number of pages and articles per pages and compares them
with the online table in anahita.
}
\examples{
\dontrun{
conn <- connect_anahita()
update_agenda_blogs(conn)
}

}
