% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-cb.R
\name{update_cb_tables}
\alias{update_cb_tables}
\title{Update CB table}
\usage{
update_cb_tables(conn, cb_token = config_get(cb_token))
}
\arguments{
\item{conn}{connection to the anahita DB}

\item{cb_token}{CrunchBase auth token (by default assumes available in config file)}
}
\description{
Updating the CrunchBase data set includes the dowlaod
of mutlitple csv files. These are stored in a temp folder.
The files are loaded and wrangled before associated tables
are pushed to the anahita DB.
}
\examples{
\dontrun{
conn <- connect_anahita()
update_cb_tables(conn1)
}

}
