% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-dms.R
\name{update_dms_tables}
\alias{update_dms_tables}
\title{Update all DMs tables}
\usage{
update_dms_tables(conn1, conn2)
}
\arguments{
\item{conn1}{connection to the data warehouse}

\item{conn2}{connection to the anahita DB}
}
\description{
Update all DMs tables
}
\examples{
\dontrun{
conn1 <- connect_dw()
conn2 <- connect_anahita()
update_dms_tables(conn1, conn2)
}

}
