% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-en.R
\name{update_en_tables}
\alias{update_en_tables}
\title{Update all EN tables}
\usage{
update_en_tables(conn1, conn2)
}
\arguments{
\item{conn1}{connection to the data warehouse}

\item{conn2}{connection to the anahita DB}
}
\description{
Update all EN tables
}
\examples{
\dontrun{
conn1 <- connect_dw()
conn2 <- connect_anahita()
update_en_tables(conn1, conn2)
}

}
