% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-gis.R
\name{update_gis_tables}
\alias{update_gis_tables}
\title{Update all GIs tables}
\usage{
update_gis_tables(conn1, conn2)
}
\arguments{
\item{conn1}{connection to the data warehouse}

\item{conn2}{connection to the anahita DB}
}
\description{
Update all GIs tables
}
\examples{
\dontrun{
conn1 <- connect_dw()
conn2 <- connect_anahita()
update_gis_tables(conn1, conn2)
}

}
