% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-master.R
\name{update_master_tables}
\alias{update_master_tables}
\title{Matching Procedure}
\usage{
update_master_tables(conn)
}
\arguments{
\item{conn}{connection to the anahita DB}
}
\description{
The company matching procedure is the backbone of the
company intelligence we intend to provide. The procedure
is intended to involve to progressively improve the matching
results. To this date, it is a very conservative approach that
uses exact matches on website URLs, company record names and
company legal names.
}
\details{
Company matching procedure
}
\examples{
\dontrun{
conn <- connect_anahita()
match_all(conn)
}

}
