% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-mixpanel.R
\name{update_mixpanel_and_views}
\alias{update_mixpanel_and_views}
\title{Update mixpanel events and views}
\usage{
update_mixpanel_and_views(
  time_since = Sys.Date() - 1,
  time_to = Sys.Date() - 1,
  dev = config_get(deploy_status),
  key_sry = config_get(mp_si),
  key_tpk = config_get(mp_dm)
)
}
\arguments{
\item{time_since}{starting time (default is yesterday)}

\item{time_to}{ending time (default is yesterday)}

\item{dev}{deployment status read from config (dev will not right into DBs)}

\item{key_sry}{mixpanel serendipidity API secret}

\item{key_tpk}{mixpanel toplink API secret}
}
\value{
a list of tibbles
}
\description{
This daily automation crawls event data from the serendipity and
toplink mixpanel accounts, append to existing tables and
updates a number of daily views, mainly around tracking conversions
and dedicated events in the platform.
}
\examples{
\dontrun{
update_mixpanel_and_views()
}

}
