% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-nr2.R
\name{update_nr2_tables}
\alias{update_nr2_tables}
\title{Update all Nr2 tables}
\usage{
update_nr2_tables(conn)
}
\arguments{
\item{conn}{connection to anahita DB}
}
\description{
The company Nr2 provides us with a list of potential
leads for NC engagement in China and South Koarea. The list
is refreshed on a monthly basis more or less. The authentication
must be refreshed every 7 days.
}
\examples{
\dontrun{
conn <- connect_anahita()
update_nr2_tables()
}

}
