% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-sf.R
\name{update_sf_tables}
\alias{update_sf_tables}
\title{Update SF table}
\usage{
update_sf_tables(conn1, conn2)
}
\arguments{
\item{conn1}{connection to the data warehouse}

\item{conn2}{connection to the anahita DB}
}
\description{
Update SF table
}
\examples{
\dontrun{
conn1 <- connect_dw()
conn2 <- connect_anahita()
update_sf_tables(conn1, conn2)
}

}
