% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-si-usage.R
\name{update_si_browsing}
\alias{update_si_browsing}
\title{Update Summary Stats Table}
\usage{
update_si_browsing(conn)
}
\arguments{
\item{conn}{connection to the anahita DB}
}
\description{
Update Summary Stats Table
}
\examples{
\dontrun{
conn <- connect_anahita()
update_si_browsing(conn)
}

}
