% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-si-usage.R
\name{update_si_location}
\alias{update_si_location}
\title{Update Summary Stats Table}
\usage{
update_si_location(conn, nmax = 5, time_since = NULL, time_to = NULL)
}
\arguments{
\item{conn}{connection to the anahita DB}

\item{nmax}{max numbe of cities per user (default is 5)}

\item{time_since}{starting time (default is NULL)}

\item{time_to}{ending time (default is NULL)}
}
\description{
Update Summary Stats Table
}
\examples{
\dontrun{
conn2 <- connect_anahita()
update_si_location(conn)
}

}
