% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-si-usage.R
\name{update_si_stats}
\alias{update_si_stats}
\title{Update Summary Stats Table}
\usage{
update_si_stats(conn, time_since = NULL, time_to = NULL)
}
\arguments{
\item{conn}{connection to the anahita DB}

\item{time_since}{starting time (default is NULL)}

\item{time_to}{ending time (default is NULL)}
}
\value{
list of tibbles
}
\description{
Update Summary Stats Table
}
\examples{
\dontrun{
conn <- connect_anahita()
update_si_stats(conn)
}

}
