% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-stripe.R
\name{update_stats_for_elvin}
\alias{update_stats_for_elvin}
\title{Update stripe customers}
\usage{
update_stats_for_elvin(dev = config_get(deploy_status))
}
\arguments{
\item{dev}{deployment status read from config (dev will not right into DBs)}
}
\value{
tibble
}
\description{
This feeds a DB with a single table used by Elvin to
report numbers on the Business Performance reporting side.
}
\examples{
\dontrun{
update_stats_for_elvin()
}

}
