% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-whitelisting.R
\name{update_whitelisting_tables}
\alias{update_whitelisting_tables}
\title{Update Whitelisting table}
\usage{
update_whitelisting_tables(conn1, conn2)
}
\arguments{
\item{conn1}{connection to the data warehouse}

\item{conn2}{connection to the anahita DB}
}
\description{
Update Whitelisting table
}
\examples{
\dontrun{
conn1 <- connect_dw()
conn2 <- connect_anahita()
update_whitelisting_tables(conn1, conn2)
}

}
